use size_t in solv_MD5_Update()
Instead of unsigned long.

+++ src/md5.h
@@ -17,5 +17,5 @@ typedef struct {
 } MD5_CTX;
 
 extern void solv_MD5_Init(MD5_CTX *ctx);
-extern void solv_MD5_Update(MD5_CTX *ctx, void *data, unsigned long size);
+extern void solv_MD5_Update(MD5_CTX *ctx, void *data, size_t size);
 extern void solv_MD5_Final(unsigned char *result, MD5_CTX *ctx);
